+++ src/pipe.rs
@@ -5,6 +5,7 @@
 use interchange::Responder;
 
 use crate::api::{Reply, Request};
+use crate::backend::BackendId;
 use crate::error::Error;
 use crate::types::ClientContext;
 
@@ -71,11 +72,12 @@ cfg_if::cfg_if! {
 // https://xenomai.org/documentation/xenomai-2.4/html/api/group__native__queue.html
 // https://doc.micrium.com/display/osiiidoc/Using+Message+Queues
 
-pub struct ServiceEndpoint {
+pub struct ServiceEndpoint<I: 'static> {
     pub interchange: Responder<TrussedInterchange>,
     // service (trusted) has this, not client (untrusted)
     // used among other things to namespace cryptographic material
     pub client_ctx: ClientContext,
+    pub backends: &'static [BackendId<I>],
 }
 
 // pub type ClientEndpoint = Requester<TrussedInterchange>;